Move validation settings to a platform settings tab
