+++ src/components/_pages/platform-settings/detail/index.tsx
@@ -1,10 +1,10 @@
-import CustomTable, { TableDataRow, TableHeader } from 'components/CustomTable';
+import CertificateSettings from 'components/_pages/platform-settings/certificates/CertificateSettings';
+import UtilsSettings from 'components/_pages/platform-settings/utils/UtilsSettings';
 import TabLayout from 'components/Layout/TabLayout';
 import Widget from 'components/Widget';
 import { WidgetButtonProps } from 'components/WidgetButtons';
 
 import { actions, selectors } from 'ducks/settings';
-import { actions as utilsActuatorActions, selectors as utilsActuatorSelectors } from 'ducks/utilsActuator';
 import { useCallback, useEffect, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router';
@@ -19,24 +19,15 @@ export default function PlatformSettingsDetail() {
     const platformSettings = useSelector(selectors.platformSettings);
     const isFetchingPlatform = useSelector(selectors.isFetchingPlatform);
 
-    const health = useSelector(utilsActuatorSelectors.health);
-
-    const getFreshData = useCallback(() => {
-        dispatch(utilsActuatorActions.health());
+    const getFreshPlatformSettings = useCallback(() => {
         if (!platformSettings) {
             dispatch(actions.getPlatformSettings());
         }
     }, [dispatch, platformSettings]);
 
     useEffect(() => {
-        getFreshData();
-    }, [getFreshData]);
-
-    useEffect(() => {
-        if (!platformSettings) {
-            dispatch(actions.getPlatformSettings());
-        }
-    }, [dispatch, platformSettings]);
+        getFreshPlatformSettings();
+    }, [getFreshPlatformSettings]);
 
     const onEditClick = useCallback(() => {
         navigate(`./edit`);
@@ -56,47 +47,6 @@ export default function PlatformSettingsDetail() {
         [onEditClick],
     );
 
-    const headers: TableHeader[] = useMemo(
-        () => [
-            {
-                id: 'setting',
-                content: 'Setting',
-            },
-            {
-                id: 'value',
-                content: 'Value',
-            },
-        ],
-        [],
-    );
-
-    const data: TableDataRow[] = useMemo(
-        () =>
-            !platformSettings
-                ? []
-                : [
-                      {
-                          id: 'utilsUrl',
-                          columns: [
-                              'Utils Service URL',
-                              platformSettings.utils?.utilsServiceUrl ? (
-                                  <>
-                                      {platformSettings.utils.utilsServiceUrl}&nbsp;
-                                      {health ? (
-                                          <i className="fa fa-check-circle" style={{ color: 'green' }} aria-hidden="true" />
-                                      ) : (
-                                          <i className="fa fa-exclamation-circle" style={{ color: 'red' }} aria-hidden="true" />
-                                      )}
-                                  </>
-                              ) : (
-                                  'n/a'
-                              ),
-                          ],
-                      },
-                  ],
-        [platformSettings, health],
-    );
-
     return (
         <Container className="themed-container" fluid>
             <Widget
@@ -105,17 +55,17 @@ export default function PlatformSettingsDetail() {
                 widgetLockName={LockWidgetNameEnum.PlatformSettings}
                 widgetButtons={buttons}
                 titleSize="large"
-                refreshAction={getFreshData}
+                refreshAction={getFreshPlatformSettings}
             >
                 <TabLayout
                     tabs={[
                         {
                             title: 'Utils',
-                            content: (
-                                <div style={{ paddingTop: '1.5em', paddingBottom: '1.5em' }}>
-                                    <CustomTable headers={headers} data={data} />
-                                </div>
-                            ),
+                            content: <UtilsSettings platformSettings={platformSettings} />,
+                        },
+                        {
+                            title: 'Certificates',
+                            content: <CertificateSettings platformSettings={platformSettings} />,
                         },
                     ]}
                 />
